+++ src/functions/greetUser.js
@@ -4,6 +4,7 @@
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
 const getGuild = require("../db/functions/guild/getGuild");
+const { nth } = require("../helpers/Util.js");
 const { Embed } = require("../classes");
 module.exports = async (member) => {
     const guildDB = await getGuild(member.guild.id);
@@ -26,13 +27,15 @@ module.exports = async (member) => {
         return "channelNotFound";
     }
     channel.sendTyping();
-    let msg = guildDB.welcomeMessage;
+    let msg = guildDB.plugins.welcome.message;
     //Replace Placeholders with their values
     msg = msg
         .replace("{mention}", `${member}`)
         .replace("{tag}", `${member.user.tag}`)
+        .replace("{username}", `${member.user.username}`)
         .replace("{server}", `${member.guild.name}`)
-        .replace("{members}", `${member.guild.memberCount}`);
+        .replace("{members}", `${member.guild.memberCount}`)
+        .replace("{members_formatted}", `${member.guild.memberCount}${nth(member.guild.memberCount)}`);
     const embed = new Embed({ color: "blue" })
         .setAuthor(
             member.user.tag,