+++ src/helpers/Util.js
@@ -51,10 +51,22 @@ const lowercaseArray = (arr) => {
         .sort();
 };
 
+//https://stackoverflow.com/a/15397495
+const nth = (d) => {
+  if (d > 3 && d < 21) return "th";
+  switch (d % 10) {
+    case 1:  return "st";
+    case 2:  return "nd";
+    case 3:  return "rd";
+    default: return "th";
+  }
+}
+
 module.exports = {
     userFromMention,
     channelIdFromMention,
     lowercaseKeys,
     lowercaseVals,
     lowercaseArray,
+    nth,
 };