+++ src/schema/guildSchema.js
@@ -17,31 +17,6 @@ const guildSchema = new mongoose.Schema({
         trim: true,
         default: "w/",
     },
-    channel: {
-        type: String,
-        required: true,
-        trim: true,
-        default: "member-log",
-    },
-    welcomeMessage: {
-        type: String,
-        required: true,
-        trim: true,
-        default:
-            "Welcome {mention} to the {server} server!\nYou are our #{members} member",
-    },
-    goodByeMessage: {
-        type: String,
-        required: true,
-        trim: true,
-        default: "Good Bye {mention}!\nWe are sad to see you go!",
-    },
-    modChannel: {
-        type: String,
-        required: true,
-        trim: true,
-        default: "mod-log",
-    },
     lang: {
         type: String,
         required: true,
@@ -49,6 +24,22 @@ const guildSchema = new mongoose.Schema({
         default: "en-US",
     },
     disabled: [String],
+    plugins: {
+        type: Object,
+        default: {
+            welcome: {
+                enabled: true,
+                message: "Welcome {mention} to the {server} server!\nYou are our #{members_formatted} member",
+                channel: "member-log",
+            },
+            goodbye: {
+                enabled: false,
+                message: "Good Bye {mention}!\nWe are sad to see you go!\nWithout you, we are {{members}} members",
+                channel: null,
+            },
+            modlogs: "mod-log",
+        },
+    },
 });
 
 const Guild = new mongoose.model("Guild", guildSchema);